requirements.txt
streamlit>=1.41.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
openpyxl>=3.1.5
lxml>=5.3.0
pandas>=2.2.0
python-dateutil>=2.9.0
reportlab>=4.1.0
Pillow>=10.0.0
bcrypt>=4.0.0
